+++ extensions/whatsapp/src/approval-reactions.ts
@@ -2,13 +2,13 @@
 import type { WAMessage } from "baileys";
 import {
   approvalReactionDecisionSetsMatch,
+  buildApprovalReactionDeliveredBindingMarker,
   createApprovalReactionTargetStore,
   listApprovalReactionBindings,
   readApprovalReactionDecisionList,
   readApprovalReactionDeliveredBinding,
   readApprovalReactionPresentationBinding,
   resolveTypedApprovalReactionTarget,
-  type ApprovalReactionDecisionBinding,
   type ApprovalReactionDeliveryBinding,
   type ApprovalReactionTargetRecord,
 } from "openclaw/plugin-sdk/approval-reaction-runtime";
@@ -32,8 +32,6 @@ const DELIVERY_BINDING_CHANNEL_DATA_KEY = "whatsappApprovalReactionBindingV1";
 type WhatsAppApprovalKind = "exec" | "plugin";
 
 type WhatsAppApprovalDeliveryBinding = ApprovalReactionDeliveryBinding;
-
-type WhatsAppApprovalReactionBinding = ApprovalReactionDecisionBinding;
 
 type WhatsAppApprovalReactionResolution = {
   approvalId: string;
@@ -122,22 +120,19 @@ function readPersistedTarget(target: unknown): WhatsAppApprovalReactionTarget | 
   if (
     !value ||
     typeof value.approvalId !== "string" ||
-    !Array.isArray(value.allowedDecisions) ||
     (value.approvalKind !== "exec" && value.approvalKind !== "plugin")
   ) {
+    return null;
+  }
+  const allowedDecisions = readApprovalReactionDecisionList(value.allowedDecisions);
+  if (!allowedDecisions) {
     return null;
   }
   return {
     approvalId: value.approvalId,
     approvalKind: value.approvalKind,
-    allowedDecisions: value.allowedDecisions,
+    allowedDecisions,
   };
-}
-
-function listWhatsAppApprovalReactionBindings(
-  allowedDecisions: readonly ExecApprovalReplyDecision[],
-): WhatsAppApprovalReactionBinding[] {
-  return listApprovalReactionBindings({ allowedDecisions });
 }
 
 const APPROVAL_ID_LINE_RE = /^\s*ID:\s*(\S(?:.*\S)?)\s*$/i;
@@ -191,11 +186,9 @@ function visibleApprovalBindingMatches(
     decisionLines.push(decisionLine);
     cursor += 1;
   }
-  const knownBindings = listWhatsAppApprovalReactionBindings([
-    "allow-once",
-    "allow-always",
-    "deny",
-  ]);
+  const knownBindings = listApprovalReactionBindings({
+    allowedDecisions: ["allow-once", "allow-always", "deny"],
+  });
   const visibleDecisions = decisionLines.map(
     (line) => knownBindings.find((entry) => `${entry.emoji} ${entry.label}` === line)?.decision,
   );
@@ -223,7 +216,7 @@ export function prepareWhatsAppApprovalPayloadForDelivery(params: {
     ...params.payload,
     channelData: {
       ...params.payload.channelData,
-      [DELIVERY_BINDING_CHANNEL_DATA_KEY]: { version: 1, ...binding },
+      [DELIVERY_BINDING_CHANNEL_DATA_KEY]: buildApprovalReactionDeliveredBindingMarker(binding),
     },
   };
 }
@@ -239,9 +232,9 @@ export function registerWhatsAppApprovalReactionTarget(params: {
 }): WhatsAppApprovalReactionTarget | null {
   const key = buildReactionTargetKey(params);
   const approvalId = params.approvalId.trim();
-  const allowedDecisions = listWhatsAppApprovalReactionBindings(params.allowedDecisions).map(
-    (binding) => binding.decision,
-  );
+  const allowedDecisions = listApprovalReactionBindings({
+    allowedDecisions: params.allowedDecisions,
+  }).map((binding) => binding.decision);
   if (
     !key ||
     !approvalId ||
