+++ src/plugin-sdk/approval-reaction-runtime.ts
@@ -23,6 +23,7 @@ import type { ReplyPayload } from "./reply-payload.js";
 export { shouldSuppressLocalNativeExecApprovalPrompt } from "./approval-native-helpers.js";
 export {
   approvalReactionDecisionSetsMatch,
+  buildApprovalReactionDeliveredBindingMarker,
   normalizeApprovalReactionDecision,
   readApprovalReactionDecisionList,
   readApprovalReactionDeliveredBinding,
